# api/userProgram/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a swag test http server.",
        "title": "USER WEB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7001",
    "basePath": "/api/v1",
    "paths": {
        "/address/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建用户收获地址",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "创建用户收获地址",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Address"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Address"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/address/getList": {
            "get": {
                "description": "获取全部地址列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取全部地址列表",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Address"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/address/save": {
            "post": {
                "description": "保存用户收获地址",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "保存用户收获地址",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Address"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Address"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/address/user/delete": {
            "post": {
                "description": "删除用户收获地址",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除用户收获地址",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/addressService.DeleteAddressModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/address/user/first": {
            "post": {
                "description": "获取用户收获地址(默认第一个)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户收获地址(默认第一个)",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Address"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/address/user/getList": {
            "get": {
                "description": "获取用户收获地址列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户收获地址列表",
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Address"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
        "/user/person/login": {
            "post": {
                "description": "微信登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "微信登陆",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "addressService.DeleteAddressModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "controller.LoginRequest": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "encryptedData": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "iv": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "rawData": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
...
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.Address": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "town": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "appKey": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "guid": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                },
                "wechatOpenId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}